add markdown to confluence script

+++ install-dot.sh
@@ -1,15 +1,27 @@
 # symbolic link
 #
-cd ~
-ln -s .dotfiles/bash_profile .bash_profile
-ln -s .dotfiles/bashrc .bashrc
 
+cd ~
+
+# backup .bash_profile
+if [ -f ~/.bash_profile ]; then
+   mv ~/.bash_profile .dotfile/backup/.bash_profile
+   ln -s .dotfiles/bash_profile .bash_profile
+fi
+
+# backup .bashrc
+if [ -f ~/.bashrc ]; then
+   mv ~/.bashrc .dotfile/backup/.bashrc
+   ln -s .dotfiles/bashrc .bashrc
+fi
+
+# backup .gitconfig
 if [ -f ~/.gitconfig ]; then
-#   mv ~/.gitconfig .gitconfig_bak
+   mv ~/.gitconfig .gitconfig_bak
    ln -s .dotfiles/gitconfig .gitconfig
 fi
 
-# brew
+# install homebrew
 # https://github.com/mxcl/homebrew/wiki/installation
 /usr/bin/ruby -e "$(curl -fsSL https://raw.github.com/gist/323731)"
 
@@ -26,3 +38,7 @@ bash < <(curl -s https://raw.github.com/wayneeseguin/rvm/master/binscripts/rvm-i
 # https://github.com/creationix/nvm
 git clone git://github.com/creationix/nvm.git ~/.nvm
 
+# dotfile auto commit
+#
+git add -A
+git commit -a -m 'installed dot files'
